requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-allauth==0.41.0
django-crispy-forms==1.9.1
django-formtools==2.2
django-import-export==2.1.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
DjangoUeditor==1.8.143
docutils==0.16
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.4
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
